requirements.txt
pandas
tqdm
torch
scikit-learn
sentence-transformers
faiss-cpu
transformers